Adds the solution for the second part of the day 16 of Advent of Code 2015.
